Add a issue command expander
The GitHub and GitLab feature to accept issue commands via the comment area by
typing '/' is quite handy. It saves time and allows for multiple commands to be
executed at the same time. It also allows the user to rely on the keyboard
instead of the mouse.

The feature itself is not implemented anywhere in Gitea. It can reuse some of
the code used to expand emojis with ':' or mentions with '@'. The command
expander however is more complicated than a trivial expander. In particular:

- **Accept arguments**
  - Some commands require arguments, e.g., a list of labels for the "label"
    command.
- **Remove text before saving**
  - The UI should remove the auto completed text from the comment before
    saving. The actions taked should be visible in the preview as well (e.g., as
    elements above the preview plane).
    TODO: What about lists, where to stop?
- **Execute the commands**
  - After saving the commands should obviously be executed.
    TODO: What about failures?
- **Might check for some integrity?***
  - An `open` and a `close` together do not make any sense.
  - Most could be resolved by only allowing a command to be present once. This
    however requires some type of state.